+++ doc/code/qp_core_operator.rst
@@ -1,0 +1,6 @@
+qp.core.operator
+================
+
+.. currentmodule:: pennylane.core.operator
+
+.. automodule:: pennylane.core.operator
