--- doc/code/qp_operation.rst
@@ -1,6 +1,0 @@
-qp.operation
-=============
-
-.. currentmodule:: pennylane.operation
-
-.. automodule:: pennylane.operation
