+++ doc/index.rst
@@ -233,13 +233,13 @@ PennyLane is **free** and **open source**, released under the Apache License, Ve
 
    code/qp_capture
    code/qp_concurrency
+   code/qp_core_operator
    code/qp_devices
    code/qp_exceptions
    code/qp_ftqc
    code/qp_gate_sets
    code/qp_measurements
    code/qp_pytrees
-   code/qp_operation
    code/qp_queuing
    code/qp_tape
    code/qp_templates_core
